Offline
Checking local comits while offline
Wrote text in Atom (appears for code only).
Saved and it populated git Desktop
This section...lower left side of git in a given branch could be notes as to why code is changed...outside of notes that would be commented in code.
Attempting to commit to TimeToCode