My Hash Algorithm
I realized that i needed to tear down my project. I was coding for a non-kernel DLL (in some parts). Tonight I will rebuild the code and load into the kernel to test.
This Hash is quick attempt to write a non-collision hash, based on the ASCII value of a character. it is more dependent on the amount of "buckets" that will be used. The fewer the more likely the collisions. But, the current algorithm accounts for upper and lower case, so it is case sensitive. It also prevents not being able to use the same letters rearranged in the same length of a string..supports permutation of a given string.