Buckets Working
I am not sure why its looks separated in this file but from line 12-80 is the current program that is working. It uses the given structure in class, prints, and deletes the dynamically created list by freeing the memory used.

I now need to add to this file, the hash table and the hash function. 6:39AM Feb 6, 2018